# apps/backend/tests/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobUploadRequest": {
        "properties": {
          "job_descriptions": {
            "description": "List of job descriptions in markdown format",
            "items": {
              "type": "string"
            },
            "title": "Job Descriptions",
            "type": "array"
          },
          "resume_id": {
            "description": "UUID reference to the resume",
            "format": "uuid",
            "title": "Resume Id",
            "type": "string"
          }
        },
        "required": [
          "job_descriptions",
          "resume_id"
        ],
        "title": "JobUploadRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Resume Matcher",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/cache/entity/{entity_type}/{entity_id}": {
      "delete": {
        "operationId": "invalidate_entity_cache_api_v1_cache_entity__entity_type___entity_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entity_type",
            "required": true,
            "schema": {
              "title": "Entity Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invalidate cache entries linked to an entity",
        "tags": [
          "v1",
          "cache"
        ]
      }
    },
    "/api/v1/cache/key/{cache_key}": {
      "delete": {
        "operationId": "invalidate_cache_key_api_v1_cache_key__cache_key__delete",
        "parameters": [
          {
            "in": "path",
            "name": "cache_key",
            "required": true,
            "schema": {
              "title": "Cache Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invalidate explicit cache key",
        "tags": [
          "v1",
          "cache"
        ]
      }
    },
...
    "/api/v1/job": {
      "get": {
        "description": "Retrieves job data from both job_model and processed_job model by job_id.\n\nArgs:\n    job_id: The ID of the job to retrieve\n\nReturns:\n    Combined data from both job and processed_job models\n\nRaises:\n    HTTPException: If the job is not found or if there's an error fetching data.",
        "operationId": "get_job_api_v1_job_get",
        "parameters": [
          {
            "description": "Job ID to fetch data for",
            "in": "query",
            "name": "job_id",
            "required": true,
            "schema": {
              "description": "Job ID to fetch data for",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get job data from both job and processed_job models",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/job/upload": {
      "post": {
        "description": "Accepts a job description as a MarkDown text and stores it in the database.",
        "operationId": "upload_job_api_v1_job_upload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "stores the job posting in the database by parsing the JD into a structured format JSON",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/jobs": {
      "get": {
        "description": "Retrieves job data from both job_model and processed_job model by job_id.\n\nArgs:\n    job_id: The ID of the job to retrieve\n\nReturns:\n    Combined data from both job and processed_job models\n\nRaises:\n    HTTPException: If the job is not found or if there's an error fetching data.",
        "operationId": "get_job_api_v1_jobs_get",
        "parameters": [
          {
            "description": "Job ID to fetch data for",
            "in": "query",
            "name": "job_id",
            "required": true,
            "schema": {
              "description": "Job ID to fetch data for",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get job data from both job and processed_job models",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/jobs/upload": {
      "post": {
        "description": "Accepts a job description as a MarkDown text and stores it in the database.",
        "operationId": "upload_job_api_v1_jobs_upload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "stores the job posting in the database by parsing the JD into a structured format JSON",
        "tags": [
          "v1"
        ]
      }
    },
...
  }
}